pyproject.toml
[project]
name = "sonar-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.64.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
]

[project.scripts]
plex = "sonar_sdk.cli:main"

[tool.setuptools]
packages = ["sonar_sdk"]

